Move WorkingDaysOfYearByMonth into Domain
